requirements.txt
spacy
scikit-learn
psycopg2-binary
python-dotenv
psycopg2